# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ApiErrorResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"StaffResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string",
						"nullable": true
					},
					"address": {
						"type": "string",
						"nullable": true
					},
					"avatar": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						]
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"username",
					"first_name",
					"last_name",
					"email",
					"phone",
					"address",
					"avatar",
					"status",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiMessageResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PaginationMeta": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"total",
					"page",
					"limit",
					"totalPages"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CustomerResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string",
						"nullable": true
					},
					"address": {
						"type": "string",
						"nullable": true
					},
					"avatar": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						]
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"username",
					"first_name",
					"last_name",
					"email",
					"phone",
					"address",
					"avatar",
					"status",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CategoryResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						]
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"status",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoryListResponse": {
				"properties": {
					"categories": {
						"items": {
							"$ref": "#/components/schemas/CategoryResponse"
						},
						"type": "array"
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"categories",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_CategoryListResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/CategoryListResponse"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_CategoryResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/CategoryResponse"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreatedCategoryResponse": {
				"properties": {
					"categoryId": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"categoryId",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_CreatedCategoryResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/CreatedCategoryResponse"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCategoryBody": {
				"properties": {
					"name": {
						"type": "string",
						"example": "Apple"
					},
					"description": {
						"type": "string",
						"example": "Apple test"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCategoryBody": {
				"properties": {
					"name": {
						"type": "string",
						"example": "Apple Updated"
					},
					"description": {
						"type": "string",
						"example": "Apple test updated"
					},
					"status": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						],
						"example": "active"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"StaffLoginResponse": {
				"properties": {
					"token": {
						"type": "string"
					},
					"staff": {
						"$ref": "#/components/schemas/StaffResponse"
					}
				},
				"required": [
					"token",
					"staff"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_StaffLoginResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/StaffLoginResponse"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginBody": {
				"properties": {
					"username": {
						"type": "string",
						"example": "admin"
					},
					"password": {
						"type": "string",
						"example": "Aa@123456"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CustomerLoginResponse": {
				"properties": {
					"token": {
						"type": "string"
					},
					"customer": {
						"$ref": "#/components/schemas/CustomerResponse"
					}
				},
				"required": [
					"token",
					"customer"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_CustomerLoginResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/CustomerLoginResponse"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "BMD Training API",
		"version": "1.0.0",
		"description": "API documentation for BMD Training backend",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/admin/categories": {
			"get": {
				"operationId": "GetAllCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiResponse_CategoryListResponse_"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Get all categories",
				"tags": [
					"Admin - Categories"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiResponse_CreatedCategoryResponse_"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Create a new category",
				"tags": [
					"Admin - Categories"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCategoryBody"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Update a category",
				"tags": [
					"Admin - Categories"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCategoryBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Delete a category",
				"tags": [
					"Admin - Categories"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/admin/categories/{categoryId}": {
			"get": {
				"operationId": "GetCategoryById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiResponse_CategoryResponse_"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Get a category by ID",
				"tags": [
					"Admin - Categories"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/admin/categories/restore": {
			"put": {
				"operationId": "RestoreCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Restore a deleted category",
				"tags": [
					"Admin - Categories"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiResponse_StaffLoginResponse_"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Staff login",
				"tags": [
					"Auth - Login & Logout"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginBody"
							}
						}
					}
				}
			}
		},
		"/api/auth/client/login": {
			"post": {
				"operationId": "LoginCustomer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiResponse_CustomerLoginResponse_"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Customer login",
				"tags": [
					"Auth - Login & Logout"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginBody"
							}
						}
					}
				}
			}
		},
		"/api/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Logout authenticated user",
				"tags": [
					"Auth - Login & Logout"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}